Rename utils package to client and update references.
