+++ cli/commands/state.go
@@ -2,7 +2,7 @@ package commands
 
 import (
 	"fmt"
-	"github.com/mesosphere/dcos-commons/cli/utils"
+	"github.com/mesosphere/dcos-commons/cli/client"
 	"gopkg.in/alecthomas/kingpin.v2"
 )
 
@@ -13,19 +13,19 @@ type StateHandler struct {
 }
 
 func (cmd *StateHandler) RunFrameworkId(c *kingpin.ParseContext) error {
-	utils.PrintJSON(utils.HTTPGet("v1/state/frameworkId"))
+	client.PrintJSON(client.HTTPGet("v1/state/frameworkId"))
 	return nil
 }
 func (cmd *StateHandler) RunProperties(c *kingpin.ParseContext) error {
-	utils.PrintJSON(utils.HTTPGet("v1/state/properties"))
+	client.PrintJSON(client.HTTPGet("v1/state/properties"))
 	return nil
 }
 func (cmd *StateHandler) RunProperty(c *kingpin.ParseContext) error {
-	utils.PrintJSON(utils.HTTPGet(fmt.Sprintf("v1/state/properties/%s", cmd.PropertyName)))
+	client.PrintJSON(client.HTTPGet(fmt.Sprintf("v1/state/properties/%s", cmd.PropertyName)))
 	return nil
 }
 func (cmd *StateHandler) RunRefreshCache(c *kingpin.ParseContext) error {
-	utils.PrintJSON(utils.HTTPPut("v1/state/refresh"))
+	client.PrintJSON(client.HTTPPut("v1/state/refresh"))
 	return nil
 }
